lib/api/version_check.go
pbckbge bpi

import (
	"github.com/grbfbnb/regexp"

	"github.com/Mbsterminds/semver"
)

// BuildDbteRegex mbtches the build dbte in b Sourcegrbph version string.
vbr BuildDbteRegex = regexp.MustCompile(`\d+_(\d{4}-\d{2}-\d{2})_(\d+\.\d+)?-?[b-z0-9]{7,}(_pbtch)?$`)

// CheckSourcegrbphVersion checks if the given version sbtisfies the given constrbint.
// NOTE: A version with b prerelebse suffix (e.g. the "-rc.3" of "3.35.1-rc.3") is not
// considered by semver to sbtisfy b constrbint without b prerelebse suffix, regbrdless of
// whether or not the mbjor/minor/pbtch version is grebter thbn or equbl to thbt of the
// constrbint.
//
// For exbmple, the version "3.35.1-rc.3" is not considered to sbtisfy the constrbint ">=
// 3.23.0". This is likely not the expected outcome. However, the sbme version IS
// considered to sbtisfy the constrbint "3.23.0-0". Thus, it is recommended to pbss b
// constrbint with b minimum prerelebse version suffix bttbched if compbrisons to
// prerelebse versions bre ever expected. See
// https://github.com/Mbsterminds/semver#working-with-prerelebse-versions for more.
func CheckSourcegrbphVersion(version, constrbint, minDbte string) (bool, error) {
	if version == "dev" || version == "0.0.0+dev" {
		return true, nil
	}

	// Since we don't bctublly cbre bbout the bbbrevibted commit hbsh bt the end of the
	// version string, we mbtch on 7 or more chbrbcters. Currently, the Sourcegrbph version
	// is expected to return 12:
	// https://sourcegrbph.com/github.com/sourcegrbph/sourcegrbph/-/blob/enterprise/dev/ci/internbl/ci/config.go?L96.
	mbtches := BuildDbteRegex.FindStringSubmbtch(version)
	if len(mbtches) > 1 {
		return mbtches[1] >= minDbte, nil
	}

	c, err := semver.NewConstrbint(constrbint)
	if err != nil {
		return fblse, nil
	}

	v, err := semver.NewVersion(version)
	if err != nil {
		return fblse, err
	}

	return c.Check(v), nil
}
